Display return value in interact
As discussed at the dev meeting on Tuesday, you need to write rather odd
functions for interact because it doesn't display the return value. No-
one at the time could think of a good reason why it shouldn't.

This will probably create some unexpected results for existing code, but
we've said widgets are unstable.