Better handling of nested widgets
I've run into situations where it'd be nice to start with one
widget, and use the value of that widget to dynamically generate
another widget. Unfortunately, when the outer widget changes,
it leaves all of the previous widgets it created,
which quickly becomes unmanageable.  Basically I'd like
something similar to the `clear_output` keyword,
but which also clears any widgets within the result of a widget.

(For a basic example of what I mean, check out:
egentry/Nested_Widgets.ipynb)

In the past I got around this in hackish ways, but others
might appreciate a clean fix.

Some thoughts / questions:
- Does this already exist? Am I missing it?
- Should this be triggered by a keyword (like `clear_output`)?
- Would this fix be better in `clear_output()` (although it'd need
to know which widgets to save and which to clear)
- This flickers, since it doesn't wait for a new widget to be
created. I could look into fixing that, but I'm not sure
where to start.
- Does this break anyone's own use cases? Are there behaviors
I'm not thinking about?  Will this break linked widgets?