Refactor lab extension
Make it so that we export the token/interface from the main package, and the CSS is only imported in the plugin. This ensures that downstream plugins don’t include the CSS (which they shouldn’t) and only import from the main package instead of a module deep in our package (which they should).