Fix a tricky lifecycle issue with tabs.
We were facing issues with the index updating automatically as we inserted new tabs into the tab bar, and keeping that consistent with the model’s selected index. So we unhook the tab bar index from the model index while we update tabs, and we also are a lot more careful about how tabs are updated to preserve the order of the (possibly delayed) tab contents.