Revert dropdown control to use a native select control.
Currently there are serious keyboard handling problems with the dropdown control (see https://github.com/ipython/ipywidgets/issues/1049), and a lack of documentation for how the control behaves (see https://github.com/ipython/ipywidgets/issues/1055). We revert to the native select control until we rectify both of these issues and split the UI control off to be a phosphor control which the widget wraps.