Remove the Proxy and PlaceProxy widgets.
Many uses of the Proxy widget can use a box with a single child instead. The PlaceProxy widget modifies arbitrary DOM elements on a page, which gets very tricky especially with multiple copies of the same notebook in a browser window. If there is a compelling need for something like PlaceProxy, we can spin it off into a separate widget package.