Listen to changes in the model outputs list.
This lets us modify the outputs from the kernel side and see the changes reflected in the widget.

Because the model state is really stored in two different places (the output area and the outputs attribute), we have to be careful when processing an iopub message. For efficiency, we want to both send the message out to be appended to existing outputs, and add it to the model’s output attribute without triggering a rerender. We do this by using the newMessage attribute in the set, which is checked for in the event handlers.