Correct use of `save_changes` and `touch` methods
I have been trying to `save_changes()` using the view object and getting an error. Apparently this part of the API has been changed per this discussion: 

https://github.com/jupyter-widgets/ipywidgets/issues/1052#issuecomment-274836038

I have attempted to correct the explanation for use of the `save_changes` method (it is to be used on the model, not the view) and add a line with an explanation for the view `touch` method. Hopefully everything is accurate.