Don't use all widgets by default
It rarely makes sense to want to export all widgets, as we don't have a
good way to tell which widget views should actually be inluded (any
layouts will cause widgets to appear twice, once on their own, and once
in the layout).

It is still able to include all by specifying `widgets=None`, and for
this case we now at least only include views of DOMWidgets.