Align the horizontal slider with the description and readout.
Fixes the problem noted in https://github.com/jupyter-widgets/ipywidgets/issues/1133#issuecomment-317452153

We might decide on a convention for how widgets react to having boxes bigger. This change makes the slider center in a large box, but other widgets stay at the top. We should change this to flex-start if we want the slider to stay at the top of the box.