Fix invalid readout_format in docs
The widget list page currently displays a traceback because the readout_format
for some of the sliders is set to 'i', which is not a valid d3 format (it's
also not a valid Python format).

Previously, this didn't result in a traceback because d3-format doesn't throw
an exception when the wrong type gets passed in -- it just silently converts it
to the type '', corresponding to 'either decimal or exponent notation, but trim
insignificant zeros'.