Delete reference to es6-promise.
This was causing problems with the html embed script being loaded via a script tag on a page with require defined - require complained about an anonymous module (see http://requirejs.org/docs/errors.html#mismatch). Basically, since es6-promise tries to detect if require is defined, it can mess up things if it is inside of a webpack bundle inside of a script tag.

So instead, we’ll just require that the page support promises (or include the polyfill if needed).