Remove confusing paragraph on importing requireJS
The paragraph on importing requireJS was a bit confusing, and, in my opinion, didn't add very much value to the documentation: either the user uses embed_minimal_html, in which case they need to import requireJS, or they use embed_data, in which case it is not relevant.

This commit removes it.