Use python compare in _should_send_property
Partial revert of 6445ea9c43282a081dc2da0a76e5844721685e56.
Main reason for reversion: Was detecting false changes when a CFloat
received a integer formatted value (i.e receives '1', which is not the
same as the coerced '1.0'). Python comparison will say these are equal
and prevent a resend.

Original reason for changing this code was also based on flawed logic
(no widgets are created since they are compared in JSON-able state, and
not passed through deserializers).