Make the basic phosphor css less specific
This is so that it doesn’t override widget-specific settings.

Fixes #1837

We namespace most of the phosphor css in case a different version of phosphor is on the page - our embedded bundle's phosphor would then stomp on the global phosphor's css.

However, we now make an exception for the top-level `.p-widget` css, since it is causing problems with specificity in our widget css selectors (i.e., overriding things it shouldn't). We could say that it is up to the widget manager to put this one snippet on the page, but for backwards compatibility, we just do it still. This has the possibility of stomping on possibly different phosphor css already on the page, but I think that conflict unlikely as the `.p-widget` css is very minimal.