When a view is removed, hide the container containing the rendered view and delete the model id from the view mimetype.
As a new convention, an empty string for the model id indicates that the view was explicitly removed and should be hidden.

Fixes #1845

This actually doesn’t quite remove all of the empty space in JuptyerLab because there is an extra nesting of containers for rendered outputs, but it does make it better, and does seem to remove space in the separate (simplified) output area views.