Handle kernel widget restoration when we already have some widgets created.
If there is a slow kernel startup, we might have some time in between the jlab widget manager’s constructor and the kernel’s connect signal. This means we may restore from the kernel on widget manager construction, then execute some cells, then restore again when the kernel signals it is connected. We fix two problems in this situation.

1. Since we might have a cell execution before the kernel connect signal, there may be some widgets already created. In that case, we want to avoid recreating those widgets. In particular, we should not take over their comm objects and replace their on_msg handler.

2. Previously, we were sending a comm message to request state and just listening for the next message back on the comm that was an update message. However, if the comm was already active, the next message back *might* be a normal update message, rather than a full state update, which would mess up our logic. We fix this by listening for an update comm message with a parent message header matching our sent message.

Finally, we fix one more issue - at startup, we want to restore widgets from the notebook and the kernel, but on a kernel connect signal, we only want to restore widgets from the kernel.

This still does not handle the case where we disconnect from the kernel, miss some update messages, and then reconnect. In order to handle that case, it would be good to have a way to ask an existing widget to update its own state with a full refresh from the kernel.