Change fontawesome support to work for fontawesome 4 and 5.
JupyterLab and notebook are upgrading to fontawesome 5 (with the v4 backwards compatibility shim). However, in some places, our CSS is very specific to version 4. This commit modifies these places to use syntax that should for both fontawesome 4 and 5. It involves some DOM structure changes, but that is a private implementation detail, so is backwards compatible with our public API.

This is a forward-port of https://github.com/jupyter-widgets/ipywidgets/pull/2793 and 5941cba85fb0959572d59af5301b34d89e18e653

Fixes #2794