Remove JupyterLab dependency in jupyterlab_widgets
Even though jupyterlab_widgets is useless without JupyterLab, it is only
a small Python package (that don't need JupyterLab) and some static
files installed in the right path (share/jupyter/labextensions). I would
then expect it to not depend on JupyterLab.

This will allow other widgets extension to depend on the jupyterlab_widget
package so that those widgets extensions can work on JupyterLab out of
the box, without them depending on JupyterLab.

Signed-off-by: martinRenou <martin.renou@gmail.com>