Do not suppress exceptions in Output widget context (#3208)
Output widgets, when used as context managers in a with statement
and in IPython, were suppressing exceptions (whose stacktrace would
be printed to the output widget).

This change allows Output widgets to suppress an exception when
`catch_exception` is set to False, which would make the control flow
of the code transparent and same as if output widgets were not used.

For backward compatibility reasons, the default behavior is not yet
changed at the moment of ipywidgets 8.0, i.e., the context manager of
Output widgets will consume and suppress any exceptions thrown.
However, at later versions the default behavior may change.