Make the standard controls optional for the HTML manager.
This allows the HTML manager to instead load, for example, older versions of the core controls. The idea is that the HTML Manager can load either the ipywidgets 7 or ipywidgets 8 version of core controls, to support either during a transition time.

It also allows the HTML manager to be used without loading all of the base controls, if desired, making it strictly more flexible.