Add a control channel message to request registered widgets.
This could be used to determine what widgets we can invoke from the frontend. We can also use this to query what version of the base widget is registered, which gives clues about what version of ipywidgets is installed.

To determine what version of ipywidgets is installed, this is a rather indirect route, so it's not so satisfactory for that purpose.