Updated dates for contributor list for ipywidgets 8 to include most commits in ipywidgets 8.
It appears that the previous date came from looking at `git merge-base master 7.x`, but there have been several merges from 7.x to master, so it's more appropriate to look at the first commits of `git log 7.x...master`. I picked a date where commits seemed to be increasing in that list as the start date of the 8.x development. It's not perfect, and the Github api doesn't allow us to just look at commits in a specific git range, but I suppose it's better than nothing.