Use JupyterLuminoPanelWidget for html manager output widget's lumino widget
Previously the processLuminoMessage was not getting called because Panel did not know to call it. This messed up the message propatation in the html manager output widget, with such affects as the layout widget for the output widget not getting initialized, the displayed event not being triggered, etc.

We also take the opportunity to make the JupyterLab output widget use the base JupyterLuminoPanelWidget, rather than having its own copy of the same code.

We also incorporate some fixes from the jlab output widget view into the html output widget view.

Fixes #3529