Fix unnecessary deprecation warnings when the widget class is initialized
Because deprecated attributes of the Widget class are static attributes, they will be accessed when initializing this class by traitlets. In that case, since a user did not explicitly try to use these attributes, we do not want to throw a deprecation warning. So we check if the thing calling these static properties is one of the known initialization functions in traitlets.