examples: fix whitespaces in Markdown links
I've found a bug with how a link to "output widget examples" is rendered
[here](https://ipywidgets.readthedocs.io/en/latest/examples/Widget%20List.html#output).

The root cause is that Markdown doesn't allow whitespaces in link destinations, so
either whitespaces should be replaced with `%20` for the links to work or the
entire URI should be enclosed in pointy bracketes (reference:
https://spec.commonmark.org/0.30/#example-487).

I've looked at neighbor links to see how it was done there and then decided to
go with `%20` to make it uniform.

Then I made a regexp to find all instances of such an issue:

    git grep -n -P '\[[^\]]*?\]\([^ )]*? [^)]*?\)'

This patch adds `%20` to Markdown links in `docs/source/examples/` and changes
an absolute URL pointing to https://ipywidgets.readthedocs.io/ to a
relative link in one of the cases with the issue.