fix: restoring widgets from saved state is broken since v8
In 9d999d7b waiting for sessionContext.ready was removed, this
caused _loadNotebook() not to work, because it was called before the
widget metadata was available.

This commit restores waiting for sessionContext.ready. This has
to be before _loadFromKernel(), otherwise this.kernel is not
present, causing an exeption, skipping _loadNotebook().

Fixes: https://github.com/jupyterlab/jupyterlab/issues/15361